Fix: move example definitions out of headers to avoid duplicate definitions
